+++ CHANGELOG.md
@@ -10,11 +10,13 @@
     * `cn.zhxu:xjsonkit-jackson:1.4.2`
     * `cn.zhxu:xjsonkit-snack3:1.4.2`
     * 参考：https://gitee.com/troyzhxu/xjsonkit
+  * 新增 `ListFieldConvertor` 字段转换器，可支持将 字符串字段自动转为简单 List 对象。
   * 升级 `slf4j-api` -> `2.0.6`
 
 * Bean Searcher Boot Starter
   * 新增 `bean-searcher.field-convertor.use-json` 配置项，表示是否自动添加 `JsonFieldConvertor`，默认 `true`
-
+  * 新增 `bean-searcher.field-convertor.use-list` 配置项，表示是否自动添加 `ListFieldConvertor`，默认 `true`
+  * 新增 `bean-searcher.field-convertor.list-item-separator` 配置项，用与配置如何将一个字符串分割成 `List` 字段
 
 # v4.0.0.alpha2 @ 2022-12-02